+++ src/args.rs
@@ -1,13 +1,14 @@
 //! Global configuration.
 
-use super::commands::Commands;
+use super::commands::Command;
 use super::error::Error;
 
 
 #[derive(Clone, Debug, clap::Parser)]
+#[command(disable_help_subcommand = true)]
 pub struct Args {
     #[command(subcommand)]
-    command: Commands,
+    command: Command,
 }
 
 impl Args {