+++ src/commands/mod.rs
@@ -9,7 +9,7 @@ use super::error::Error;
 
 
 #[derive(Clone, Debug, clap::Subcommand)]
-pub enum Commands {
+pub enum Command {
     /// Query the DNS.
     Query(self::query::Query),
 
@@ -17,15 +17,15 @@ pub enum Commands {
     Lookup(self::lookup::Lookup),
 
     /// Show the manual pages.
-    Man(self::help::Help),
+    Help(self::help::Help),
 }
 
-impl Commands {
+impl Command {
     pub fn execute(self) -> Result<(), Error> {
         match self {
             Self::Query(query) => query.execute(),
             Self::Lookup(lookup) => lookup.execute(),
-            Self::Man(help) => help.execute(),
+            Self::Help(help) => help.execute(),
         }
     }
 }