+++ Changelog.md
@@ -4,11 +4,22 @@
 
 Breaking changes
 
+* Renamed the `man` command to `help`. ([18])
+* The default query type for the `query` command is now `AAAA`. ([#4])
+
 New
+
+* Added a new `lookup` command. ([#10])
 
 Bug fixes
 
 Other changes
+
+* Increased minimum supported Rust version to 1.78.
+
+[#4]: https://github.com/NLnetLabs/dnsi/pull/4
+[#10]: https://github.com/NLnetLabs/dnsi/pull/10
+[#18]: https://github.com/NLnetLabs/dnsi/pull/18
 
 
 ## 0.1.0